MdeModulePkg/CapsuleRuntimeDxe: Merge changes form arm to all ARCH
BZ: https://bugzilla.tianocore.org/show_bug.cgi?id=1462

The arm ARCH has already to add a function CapsuleCacheWriteBack to
flush the cache data to DRAM. That is also required in X86 ARCH. So
merge the changes. And this function do not support in runtime phase.

Contributed-under: TianoCore Contribution Agreement 1.1
Signed-off-by: Zhichao Gao <zhichao.gao@intel.com>
Cc: Ray Ni <ray.ni@intel.com>
Cc: Liming Gao <liming.gao@intel.com>